Discord Bot/requirements.txt
discord.py
python.dotenv
Flask
tweepy